+++ cloud_build/build.R
@@ -1,0 +1,7 @@
+library(googleCloudRunner)
+
+cr_deploy_pkgdown(
+  "cloudyr/googleCloudStorageR",
+  secret = "github-ssh",
+  create_trigger = "inline"
+)