+++ NEWS.md
@@ -5,6 +5,7 @@
 * Allow GCS uploads if ACL has been set at bucket level (#111)
 * Let users set the boundary between simple and resumable upload limit via new option `option(googleCloudStorageR.upload_limit)` or `gcs_upload_set_limit()` - default is 5000000L or 5MB (#120)
 * Suppress warning for when resumable uploads are requesting the upload URL (#120)
+* Fix ability to upload meta data with objects (#108)
 
 # googleCloudStorageR 0.5.1